// src/modules/offering/income/enums/offering-income-search-type.enum.ts
export enum OfferingIncomeSearchType {
  SundayService = 'sunday_service',
  FamilyGroup = 'family_group',
  GeneralFasting = 'general_fasting',
  GeneralVigil = 'general_vigil',
  ZonalFasting = 'zonal_fasting',
  ZonalVigil = 'zonal_vigil',
  GeneralEvangelism = 'general_evangelism',
  ZonalEvangelism = 'zonal_evangelism',
  SundaySchool = 'sunday_school',
  YouthService = 'youth_service',
  TeenagerService = 'teenager_service',
  UnitedService = 'united_service',
  ZonalUnitedService = 'zonal_united_service',
  Activities = 'activities',
  ChurchGround = 'church_ground',
  Special = 'special',
  IncomeAdjustment = 'income_adjustment',
  RecordStatus = 'record_status',
}

export const OfferingIncomeSearchTypeNames: Record<
  OfferingIncomeSearchType,
  string
> = {
  [OfferingIncomeSearchType.SundayService]: 'Culto Dominical',
  [OfferingIncomeSearchType.FamilyGroup]: 'Grupo Familiar',
  [OfferingIncomeSearchType.GeneralFasting]: 'Ayuno General',
  [OfferingIncomeSearchType.GeneralVigil]: 'Vigilia General',
  [OfferingIncomeSearchType.ZonalFasting]: 'Ayuno Zonal',
  [OfferingIncomeSearchType.ZonalVigil]: 'Vigilia Zonal',
  [OfferingIncomeSearchType.GeneralEvangelism]: 'Evangelismo General',
  [OfferingIncomeSearchType.ZonalEvangelism]: 'Evangelismo Zonal',
  [OfferingIncomeSearchType.SundaySchool]: 'Escuela Dominical',
  [OfferingIncomeSearchType.YouthService]: 'Culto Jóvenes',
  [OfferingIncomeSearchType.TeenagerService]: 'Culto Adolescentes',
  [OfferingIncomeSearchType.UnitedService]: 'Culto Unido General',
  [OfferingIncomeSearchType.ZonalUnitedService]: 'Culto Unido Zonal',
  [OfferingIncomeSearchType.Activities]: 'Actividades',
  [OfferingIncomeSearchType.ChurchGround]: 'Terreno Iglesia',
  [OfferingIncomeSearchType.Special]: 'Ofrendas Especial',
  [OfferingIncomeSearchType.IncomeAdjustment]: 'Ajustes por Ingreso',
  [OfferingIncomeSearchType.RecordStatus]: 'Estado de Registro',
};